Implementación de confirmación de eliminación en el componente de listado de cursos. Se añadió el servicio SweetAlert para mostrar un diálogo de confirmación antes de proceder con la eliminación de un curso. Se actualizaron los métodos correspondientes para manejar la lógica de confirmación y se mejoró la experiencia del usuario al eliminar registros.

+++ frontend/src/app/components/curso/listar/listar.component.ts
@@ -2,6 +2,7 @@ import { Component, OnInit } from '@angular/core';
 import { ToastrService } from 'ngx-toastr';
 import { Curso } from 'src/app/interfaces/curso';
 import { CursoService } from 'src/app/services/curso';
+import { SweetAlertService } from 'src/app/services/sweet-alert.service';
 
 @Component({
   selector: 'app-listar',
@@ -14,6 +15,7 @@ export class ListarComponent implements OnInit {
 
   constructor(
     private _cursoService: CursoService,
+    private _sweetAlert: SweetAlertService,
     private toastr: ToastrService,
   ) { }
 
@@ -31,12 +33,22 @@ export class ListarComponent implements OnInit {
     })
   }
 
-  Delete(id: number) {
-    this.loading = true;
-    this._cursoService.delete(id).subscribe(() => {
-      this.GetList();
-      this.toastr.error(`El Registro fue elimado con exito`, 'Curso Eliminado');
-    })
+  async Delete(id: number) {
+    const result = await this._sweetAlert.mostrarConfirmacion(
+      "¿Estás seguro?",
+      "¡No podrás revertir esta acción!"
+    );
+
+    // Verificar si el usuario confirmó la eliminación
+    if (result.isConfirmed) {
+      // Si el usuario confirma, procedemos con la eliminación
+      this.loading = true;
+      this._cursoService.delete(id).subscribe(() => {
+        this.GetList();
+        this.toastr.error(`El Registro fue eliminado con éxito`, 'Curso Eliminado');
+      });
+    }
+
   }
 
 }
